+++ examples/src/app/angular4-paystack/paystack-options.ts
@@ -12,4 +12,5 @@ export interface PaystackOptions {
   bearer: string;
   callback: (reference: string) => void ;
   onClose:  () => void;
+  container?: string
 }